merge from master

+++ service/src/main/java/cz/fi/muni/pa165/service/OrderServiceImpl.java
@@ -31,29 +31,45 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     public void create(Order order, String userLogin) {
+        if (order == null) {
+            throw new IllegalArgumentException("Order cannot be null.");
+        }
+        if (order.getServices().isEmpty() && order.getTires().isEmpty()) {
+            throw new IllegalArgumentException("Cannot create empty order.");
+        }
         User user = userDao.getUserByLogin(userLogin);
         order.setUser(user);
-        orderDao.create(order);
-
-        /*
-        List<Order> orders = null;
-        orders = user.getOrders();
-        orders.add(order);
-        user.setOrders(orders);
-        userDao.updateUser(user);
-        
-         */
-
+        try {
+            orderDao.create(order);
+        } catch (DataAccessException ex) {
+            throw new RuntimeException(ex);
+        }
     }
 
     @Override
     public void remove(Order order) {
-        orderDao.remove(order);
+        if (order == null) {
+            throw new IllegalArgumentException("Order cannot be null");
+        }
+        try {
+            orderDao.remove(order);
+        } catch (DataAccessException ex) {
+            throw new RuntimeException(ex);
+        }
     }
 
     @Override
     public Order findById(Long id) {
-        return orderDao.findById(id);
+        if (id == null) {
+            throw new IllegalArgumentException("Order id cannot be null");
+        }
+        Order order = null;
+        try {
+            order = orderDao.findById(id);
+        } catch (DataAccessException ex) {
+            throw new RuntimeException(ex);
+        }
+        return order;
     }
 
     @Override
@@ -61,7 +77,7 @@ public class OrderServiceImpl implements OrderService {
         List<Order> orders = null;
         try {
             orders = orderDao.findAll();
-        } catch (DataAccessException ex){
+        } catch (DataAccessException ex) {
             throw new RuntimeException(ex);
         }
         return orders;
@@ -69,37 +85,64 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     public List<Order> getOrdersByUser(User user) {
+        if (user == null) {
+            throw new IllegalArgumentException("User cannot be null");
+        }
         return orderDao.findByUser(user);
     }
 
     @Override
     public void confirm(Order order) {
+        if (order == null) {
+            throw new IllegalArgumentException("Order id cannot be null");
+        }
         order.setState(OrderState.CONFIRMED);
     }
 
     @Override
     public void finish(Order order) {
+        if (order == null) {
+            throw new IllegalArgumentException("Order id cannot be null");
+        }
         order.setState(OrderState.DONE);
     }
 
     @Override
     public void addTireToOrder(Long orderId, Long tireId) {
+        if (orderId == null) {
+            throw new IllegalArgumentException("Order id cannot be null");
+        }
+        if (tireId == null) {
+            throw new IllegalArgumentException("Service id cannot be null");
+        }
         orderDao.addTire(orderId, tireId);
     }
 
     @Override
     public void addServiceToOrder(Long orderId, Long serviceId) {
+        if (orderId == null) {
+            throw new IllegalArgumentException("Order id cannot be null");
+        }
+        if (serviceId == null) {
+            throw new IllegalArgumentException("Service id cannot be null");
+        }
         orderDao.addService(orderId, serviceId);
     }
 
     @Override
     public Car getOrderCar(Long orderId) {
+        if (orderId == null) {
+            throw new IllegalArgumentException("Order id cannot be null");
+        }
         Order order = orderDao.findById(orderId);
         return carDao.findById(order.getCar().getId());
     }
 
     @Override
     public void cancel(Order order) {
+        if (order == null) {
+            throw new IllegalArgumentException("Order id cannot be null");
+        }
         order.setState(OrderState.CANCELED);
     }
 }
