+++ service/src/main/java/cz/fi/muni/pa165/service/ServiceServiceImpl.java
@@ -1,11 +1,9 @@
 package cz.fi.muni.pa165.service;
 
 import cz.fi.muni.pa165.dao.ServiceDao;
-import cz.fi.muni.pa165.entity.Car;
 import cz.fi.muni.pa165.entity.Service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
-
 
 import java.util.List;
 /**
@@ -19,6 +17,10 @@ public class ServiceServiceImpl implements ServiceService {
 
     @Override
     public Service findById(Long id) {
+        if(id == null){
+            throw new IllegalArgumentException("Service ID cannot be null");
+        }
+
         Service service = null;
 
         try {
@@ -41,6 +43,11 @@ public class ServiceServiceImpl implements ServiceService {
     }
     @Override
     public void create(Service service) {
+
+        if(service == null){
+            throw new IllegalArgumentException("Service object cannot be null");
+        }
+
         try {
             serviceDao.create(service);
         }catch (DataAccessException ex){
@@ -50,6 +57,9 @@ public class ServiceServiceImpl implements ServiceService {
 
     @Override
     public void remove(Service service) {
+        if(service == null){
+            throw new IllegalArgumentException("Service object cannot be null");
+        }
         try {
             serviceDao.remove(service);
         }catch (DataAccessException ex){
